+++ src/mathematics/divide.c
@@ -2,13 +2,13 @@
 
 void divide()
 {
-    float input1, input2;
+    long double input1, input2;
 
     printf("\nEnter starting number: ");
-    scanf("%f", &input1);
+    scanf("%Lf", &input1);
 
     printf("\nEnter number to divide by: ");
-    scanf("%f", &input2);
+    scanf("%Lf", &input2);
 
-    printf("\nResult: %f\n", input1 / input2);
+    printf("\nResult: %.2Lf\n", input1 / input2);
 }
