+++ src/mathematics/rectangle_perimeter.c
@@ -2,13 +2,13 @@
 
 void rectanglePerimeter()
 {
-    float side1, side2;
+    long double side1, side2;
 
     printf("\nEnter the length of one side: ");
-    scanf("%f", &side1);
+    scanf("%Lf", &side1);
 
     printf("\nEnter the length of another side: ");
-    scanf("%f", &side2);
+    scanf("%Lf", &side2);
 
-    printf("\nResult: %f\n", (side1 * 2) + (side2 * 2));
+    printf("\nResult: %.2Lf\n", (side1 * 2) + (side2 * 2));
 }
